Implement integration test for reminder registration and reminder callback
The following has been implemented:
- Base class `ActorIntegrationTest` which contains `GetActorService`
which creates an `ActorService` objects and mocks its dependencies. It
uses `NullActorStateProvider` as a state provider which stores data
in-memory and does not have any dependencies to the SF runtime. This
makes it suitable for implementing simple integration tests.
- Test `ReminderParametersPersistBetweenRegisteringAndReceivingReminder`
which creates an `ActorService` which hosts actors of type
`TestableActor`. The test registers a reminder on one actor, checks the
return value, waits for the reminder to fire and finally checks whether
the callback was executed successfully.

Changes to GitHub actions:
- The "Build Test" action has been modified by adding "Uninstall SF SDK
assemblies from GAC". This step removes
`Microsoft.ServiceFabric.Data.Interfaces` and
`Microsoft.ServiceFabric.Data.Interfaces.V2` from the GAC on the runner.
`Microsoft.ServiceFabric.Actors.ItegrationTests` targeting .NET
Framework 4.6 references `Microsoft.ServiceFabric.Data` version 9
package which contains `Microsoft.ServiceFabric.Data.Interfaces.dll`
version 5.0. The runner has
`Microsoft.ServiceFabric.Data.Interfaces.dll` version 5.0 installed in
the GAC, but this assembly corresponds to `Microsoft.ServiceFabric.Data`
version 8. In this case, the .NET runtime chooses to load the assembly
found in the GAC, which causes a `System.MissingMethodException` at
runtime because the wrong version of the assembly is being used. These
changes to the GitHub action are a temporary solution until assembly
versions are changed to match the version of the SDK they belong to.