Enable MeterProvider creation without ServiceContext and minor changes (#481)
In order to delete old telemetry API, we need to replace its usage in IS
(Infrastructure Service) with the new API.
An issue we found is that Meters in IS are currently created outside of
the Service context, before service is instantiated.

We need to keep the old functionality, so we are enabling creation of
MeterProviders with null ServiceContext. MeterProviders created in this
way will create Meters that emit not system dimensions based on
ServiceContext. Note that system dimensions emitted from the native site
(nodeId, runtimeVersion etc.) will still be emitted as normal, this only
applies to system dimensions read from ServiceContext.

In addition to this we are making Diatgnostics internals visible to
relevant IS projects.

---------