Refactor usage of Performance Counters in ServicesRemoting and introduce abstractions needed for Telemetry (#441)
**Motivation**
- We are not intending to change the behavior of the system in this PR.
- We want to refactor the way Performance Counters are emitted and to
encapsulate it in an easily extendable abstraction, so that we can
easily integrate telemetry in the future.
- In this PR we are trying to stray away from the existing pattern used
for DiagnosticsManager in Actors project, but rather use Composite
pattern.
- Use TDD for the development of this feature.

**Note**
- Before this PR is integrated in SF, we need to follow the process and
publish new public NuGet packages, this is needed because we lacked
InternalsVisibleTo for certain classes used in the ServicesRemoting
projects. These classes were needed for testing, so update is necessary.