Remove unused FabricTransport code.
Most of the code in the original Microsoft.ServiceFabric.FabricTransport.dll
is based on the native FabricServiceCommunications.dll. This managed assembly
only has InternalsVisibleTo the Microsoft.ServiceFabric.Services.Remoting.dll
also built out of this repo. Now that both managed DLLs live here, we can
can safely remove the unused internal APIs with the help of the compiler.